### 2026-07-16 #10987 @editorialbot
[PRE REVIEW]: nsEVDx: A Python library for modeling non-stationary extreme value distributions
**Submitting author:** @Nischalcs50 (<a href="http://orcid.org/0009-0004-3187-4920">Nischal Kafle</a>)
**Repository:** https://github.com/Nischalcs50/nsEVDx
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.4
