### 2026-08-18 #10987 @manunicholasjacob
@boisgera happy to take this one if you still need reviewers.

I'd come at it from the software and performance side. The MCMC engine is where I'd spend most of my time. Three samplers with very different per-step costs, all hand-rolled in NumPy, is a place where documented behaviour and measured behaviour can drift apart without anyone noticing. The acceptance rate diagnostics give me something to check them against. Install, test suite and docs pass alongside that.

Most of my measurement work is on latency and energy distributions on edge hardware. The tail is normally what breaks a deployment, not the mean, so a library aimed at non-stationary tails interests me for reasons that have nothing to do with rainfall or stock markets.

I maintain two open benchmarking tools, llama-roofline (10.5281/zenodo.21842493) and ml-systems-lab (10.5281/zenodo.21867055). My habit is to measure timing claims instead of taking them on trust, and to say so when I can't verify one on the hardware I have.

One scheduling note so it isn't a surprise later. I'm on the USENIX ATC 2026 Artifact Evaluation Committee from 22 September to 14 October, so I'll be slower through that stretch. It doesn't stop me taking this on and I'll keep you posted.

I'm in the reviewer database as @manunicholasjacob. ORCID 0009-0007-6589-6572.
