### 2026-08-23 Nischalcs50/nsEVDx#6 @manunicholasjacob
[JOSS Review] Sign error in the MALA acceptance ratio
Part of the JOSS review at openjournals/joss-reviews#11187.

I found what appears to be a sign error in the MALA acceptance ratio.

In `_Mala_1chain` (`nsEVDx/evd_model.py`, line 569 on current master), the
current expression effectively adds the forward proposal density instead of
subtracting it. The Metropolis-Hastings ratio should subtract both the current
log posterior and the forward proposal density. Because of the current
parenthesization, the forward density ends up being added instead.

I tested this using a simple three-dimensional N(0, 1) target by overriding the
posterior hooks on a model instance. With the implementation as it currently
stands, the sampled standard deviations were approximately 0.60 rather than the
expected 1.00. Changing only the acceptance-ratio line recovered standard
deviations of approximately 1.00. The same pattern holds at other settings I
tried; with a target standard deviation of 2, the sampler produced
approximately 1.22 as shipped and approximately 2.01 with the corrected ratio.

This matters because the sampler is no longer targeting the intended posterior,
and in this case the effect is to underestimate posterior spread and produce
overly narrow credible intervals.

The fix itself appears to be a one-line change. I have also prepared a
regression test based on a known Gaussian target so that the sampler is tested
on whether it samples the correct distribution, rather than only whether it
runs successfully and returns outputs of the expected shape.

I have a reproduction script and measurements available if helpful.
