### 2026-08-23 Nischalcs50/nsEVDx#7 @manunicholasjacob
[JOSS Review] Reported acceptance rates include burn-in acceptances
Part of the JOSS review at openjournals/joss-reviews#11187.

I noticed that `_Mala_1chain` and `_RandWalk_1chain` count acceptances
throughout both burn-in and sampling, but report the acceptance rate by dividing
only by `num_samples`.

As a result, any nonzero burn-in inflates the reported acceptance rate and can
even produce a value above 1.0.

For example, in a stationary GEV run with equal numbers of burn-in and retained
samples, the reported acceptance rate was 105.6%, while the rate measured
directly from the retained chain was about 53%. The software then warned that
the reported rate was outside the recommended range even though the actual
production acceptance rate was within it. With burn-in set to twice the number
of retained samples, the reported rate was almost exactly three times the
measured one, as the mismatched denominator predicts.

This could affect users tuning proposal widths or step sizes, since they may be
making those decisions based on an inflated number.

HMC already calculates its acceptance rate over the production sampling loop
only. It would make sense for the other samplers to follow the same convention.

I have prepared regression tests for both samplers.
