### 2026-08-23 Nischalcs50/nsEVDx#8 @manunicholasjacob
[JOSS Review] HMC mass matrix appears to use the inverse convention
Part of the JOSS review at openjournals/joss-reviews#11187.

I think there may be an issue with the convention used when estimating the HMC
mass matrix.

Under the conventions used by the sampler, the momentum distribution, kinetic
energy, and leapfrog position update imply that the diagonal mass matrix should
scale with posterior precision (Neal 2011, section 4.1; Betancourt 2017, which
the paper cites). The current implementation instead estimates it from posterior
variance in both the initial estimate and the warmup update
(`nsEVDx/hmc_engine.py`, lines 341 and 421 on current master).

This has the opposite effect from the intended preconditioning: dimensions with
wider posterior distributions receive smaller effective position steps.

I tested this directly using an independent Gaussian target with very different
scales across the two dimensions. With the current convention, the wide
dimension mixed extremely poorly. In a run with 4,000 retained samples, it had
an effective sample size of about 3 and a sampled standard deviation of 14.8 for
a target with a standard deviation of 50. Using the precision convention instead
gave effective sample sizes near the full sample count and recovered the target
much more closely. The same pattern held when I varied the scale ratio (10:1
and 100:1).

I saw a similar pattern in an end-to-end stationary GEV fit, where HMC produced
a much lower minimum ESS than random-walk Metropolis despite requiring
substantially more gradient evaluations.

The suggested fix is to invert the current estimates so that the initialization
uses curvature and the warmup update uses inverse variance.

I can share the benchmark script if useful.
