### 2026-08-23 Nischalcs50/nsEVDx#9 @manunicholasjacob
[JOSS Review] GPD likelihood does not enforce the lower support bound
Part of the JOSS review at openjournals/joss-reviews#11187.

While testing the samplers on a GPD workflow, I found that the non-stationary
GPD likelihood does not fully enforce the distribution's support.

The current GPD branch (in `neg_log_likelihood_ns`, `nsEVDx/utils.py`) checks
whether `1 + xi*z > 0`, but it does not also reject `z < 0`. For a GPD,
observations must be at or above the location parameter. As things stand,
points below the location can pass the existing check and still receive a
finite likelihood. For positive shape values, the likelihood can even reward
moving the location parameter above some of the data.

I tested this with 300 simulated GPD observations generated with shape 0.2,
location 0, and scale 5. At the true parameters, the package's negative
log-likelihood was 844.545 and matched the corresponding SciPy calculation to
all printed digits.

However, moving the location to 1.0 placed roughly 50 observations below the
support boundary. At that point, the true likelihood should be zero, with
SciPy returning -inf for the log-likelihood. The package instead returned an
NLL of 783.55, which is better than the value at the true parameters.

This also shows up in fitting. In a random-walk fit of a stationary GPD with a
free location parameter, the chain drifted toward a posterior centered around
a location of 2.9 and a scale of 0.37, despite the data being generated with a
location of 0 and a scale of 5.

The same missing support check is present in `_grad_nll_gpd`, so this is not
limited to random-walk sampling. MALA and HMC are affected as well because
they rely on the same underlying likelihood and gradient logic.

The GEV branch does not appear to have this problem because its support
condition is already captured by the `v > 0` check.

The fix should be relatively straightforward: reject `z < 0` in the GPD
likelihood and gradient branches, alongside the existing support checks. A
regression test comparing the NLL at the true parameters with one where the
location is moved above part of the data would help prevent this from
recurring.

This behavior is present in both the released version and the current master
branch. I have a short reproduction script available if helpful.
