### 2026-08-23 Nischalcs50/nsEVDx#10 @manunicholasjacob
[JOSS Review] Released package differs materially from the current master branch
Part of the JOSS review at openjournals/joss-reviews#11187.

The currently released package differs in several important ways from the
current `master` branch, including sampler issues that have already been fixed
in the repository.

This means that a user installing the package from PyPI today may get different
and potentially incorrect behavior from someone installing directly from the
current source.

The released version still contains an HMC double acceptance/rejection issue
that has since been fixed on `master`. It also contains a stationary-scale
gradient issue that has been corrected in the repository. In addition, the
package version reported by `nsEVDx.__version__` does not match the released
package metadata.

Since the JOSS review process will eventually require a tagged release anyway, I
think the main thing here is to make sure a new release is cut once the
outstanding review issues have been resolved and that the version metadata is
updated at the same time.

This would ensure that the reviewed version, the repository, and the package
users install are all aligned.
