### 2026-08-23 Nischalcs50/nsEVDx#11 @manunicholasjacob
[JOSS Review] A few API contracts and docstrings have drifted from the implementation
Part of the JOSS review at openjournals/joss-reviews#11187.

I found a few smaller API and documentation inconsistencies that seem worth
cleaning up.

First, `dist` is documented and accepted as either a string or a SciPy
distribution object, but parts of the frequentist path assume it always has a
`.name` attribute. As a result, using a string such as `"genextreme"` can cause
`suggest_bounds` and the frequentist workflow to fail with an `AttributeError`,
even though the Bayesian paths accept the same input. (The same pattern appears
in `utils.py`, lines 519 and 522.)

Second, `frequentist_nsEVD` is annotated and documented as returning two values,
but currently returns only the fitted parameters. Code following the documented
interface and trying to unpack both values will therefore fail.

There are also a couple of stale internal docstrings. `_hmc_step` still says
that it performs the Metropolis accept/reject step, although that logic was
moved to the callers during the recent restructuring. Similarly,
`_numerical_grad_log_posterior` only computes the likelihood component of the
gradient, with the prior gradient added elsewhere, so its current name and
documentation overstate what it returns.

None of these look like major changes individually, but together they create
unnecessary confusion between the documented API and the actual behavior.
