### 2026-08-23 Nischalcs50/nsEVDx#12 @manunicholasjacob
[JOSS Review] Paper and reference cleanup
Part of the JOSS review at openjournals/joss-reviews#11187.

I found a few small paper and reference issues that should be straightforward to
clean up.

The editorial check for a "State of the field" section appears to be failing
because the paper currently calls the section "Statement of the Field." The
relevant content is already there, so this looks like a heading-name issue
rather than missing material.

There are also a few reference problems. One citation still has a placeholder
DOI, another includes the full `https://doi.org/` prefix where the reference
checker expects the bare DOI, and one DOI for a forthcoming paper does not
currently resolve.

The second author's ORCID field is also empty, and I noticed several small typos
throughout the paper.

One optional point concerns the download claim. The statement about more than
3,200 downloads may be accurate depending on whether mirror traffic is included,
but it would be stronger to identify the counter being used and the date the
figure was checked.

These are mostly editorial fixes, but addressing them should help the paper pass
the automated checks and make the metadata more durable.
