### 2026-08-23 Nischalcs50/nsEVDx#13 @manunicholasjacob
[JOSS Review] Default priors and quick-start tempering may be misleading for new users
Part of the JOSS review at openjournals/joss-reviews#11187.

I have two related points about the defaults that new users are likely to
encounter first.

The default prior for the stationary scale parameter uses a fixed width of 0.3
in the original data units. This means the same prior can be relatively weak for
one dataset and extremely informative for another simply because the data are
measured on a different scale or in different units.

For quantities such as rainfall or discharge, where the scale parameter can be
much larger than 1, this fixed width may have a substantial influence on the
posterior. A width that scales with the sample standard deviation would behave
more consistently across datasets. At minimum, I think the behavior of the
default should be documented clearly, especially since the defaults are likely
to be what many new users try first.

The README quick start also runs HMC with `T = 1.25`. Since this samples a
tempered, flattened distribution rather than the actual posterior, the posterior
summaries produced by the first example are intentionally wider than the
untempered result.

I would suggest using `T = 1.0` in the quick start and linking to the tempering
functionality separately for users who want to use it.

I also noticed a small typo in the corresponding section heading/comment.
