### 2026-08-24 Nischalcs50/nsEVDx#17 @dosvk
[JOSS Review] README Quick Start fails on the released package (v0.2.4): MH_Hmc return signature mismatch
Part of the JOSS review at openjournals/joss-reviews#11187.

Checklist item: `Example usage`

## Summary

I tried the README Quick Start against the version under review (`pip install nsEVDx`, which gets v0.2.4) and it crashes with a `ValueError`. Annoyingly, the sampling itself finishes fine, it dies right after when unpacking the sampler's return value. So as a first-time user you wait for the whole chain to run and then lose the results.

## Environment

- macOS (arm64), Python 3.12.13, fresh venv
- `pip install nsEVDx` (resolves to 0.2.4 on PyPI)
- README Quick Start copied as-is; the only thing I added was `matplotlib.use("Agg")`
  so it runs headless

## What happens

```
HMC Chain 1: 100%|##########| 2000/2000 [00:03<00:00, 570.21sample/s]
Traceback (most recent call last):
  File "quickstart_readme.py", line 38, in <module>
    samples, a_rate = sampler.MH_Hmc(
    ^^^^^^^^^^^^^^^
ValueError: too many values to unpack (expected 2)
```

## Why

The README at v0.2.4 unpacks two values:

```python
samples, a_rate = sampler.MH_Hmc(...)   # README, ~line 112
```

but the shipped code returns four ([hmc_engine.py#L521 at the tag](https://github.com/Nischalcs50/nsEVDx/blob/v0.2.4/nsEVDx/hmc_engine.py#L521)):

```python
return samples, a_rate, step_size, divergences
```

## Note on master

I checked master before filing this. Installed from source, ran its Quick Start end-to-end: it completes fine (posterior summaries, R-hats and all). There the example reads `results = sampler.MH_Hmc(...)` and the sampler returns a dict, so README and code agree. It's basically a user-visible case of what #10 already describes: `pip install` today + the released README = crash, while master doesn't have this mismatch.

## Suggested fix

Since JOSS reviews (and archives) a tagged release, I think the real fix is cutting a new release where the README example and the shipped API actually match. Settling on a stable return contract (the dict on master looks reasonable) would also stop the README and code drifting apart again.

Happy to re-run the Quick Start against a new tag once it's up.
