### 2026-08-24 Nischalcs50/nsEVDx#18 @dosvk
[JOSS Review] Plotting tests block on figure windows without a headless backend
Part of the JOSS review at openjournals/joss-reviews#11187.

Checklist item: `Automated tests`

A nit probably.

Running `python -m pytest tests/` on my Mac (default matplotlib backend), the plotting tests open figure windows one after another and the run blocks until each window is closed by hand — the first one stops the suite at `tests/test_utils.py::test_plot_trace` around 93%. If nobody is at the screen (ssh, tmux, walked away), it just sits there, which is how I first hit it. With `MPLBACKEND=Agg python -m pytest tests/` the whole suite passes unattended, 110/110 in about 18 seconds with 97% coverage, which is genuinely good for a research package.

CI doesn't see any of this because the runners are headless. Cause is that the plotting helpers call `plt.show()` directly (utils.py lines 768 and 815 at the v0.2.4 tag) and the plot tests go through them.

`Note`: This is still the case on current master too

Two easy options: force Agg in `tests/conftest.py` (`matplotlib.use("Agg")` before anything imports pyplot), or give the plotting functions a `show=False` / return-the-figure path instead of calling `plt.show()` themselves. The second also helps anyone using the library from scripts or notebooks who wants to control the figure lifecycle.
