### 2026-08-24 Nischalcs50/nsEVDx#19 @dosvk
[JOSS Review] Bare pytest fails on a fresh clone, pytest-cov missing from requirements.txt
Part of the JOSS review at openjournals/joss-reviews#11187.

Checklist item: `Automated tests`

Small one. On a fresh clone, `python -m pytest tests/` errors out before collecting anything:

```
ERROR: usage: python -m pytest [options] [file_or_dir] [file_or_dir] [...]
python -m pytest: error: unrecognized arguments: --cov=nsEVDx --cov-report=term-missing
```

pyproject.toml sets `addopts = "--cov=nsEVDx --cov-report=term-missing"`, so pytest demands pytest-cov, but requirements.txt lists pytest without pytest-cov. Same on current master, so flagging it separately from the release fix. Either add pytest-cov to requirements.txt or move the coverage flags out of addopts (into CI or a make target), whichever you prefer. Once pytest-cov is installed the suite runs, with `MPLBACKEND=Agg` it's 110/110 (see #18 for the default-backend hang though).
