### 2026-08-24 #11187 @dosvk
A few paper-level points to go with the code issues I filed (Nischalcs50/nsEVDx#17 - Nischalcs50/nsEVDx#19).

State of the field: the statement of need compares mostly against R packages (ismev, extRemes, climextRemes, NSGEV) and the general PPLs (PyMC, Stan), but it doesn't mention the existing Python EVA packages. [pyextremes](https://pypi.org/project/pyextremes/) in particular does block-maxima and POT for GEV/GPD, MLE and MCMC-based Bayesian inference, return levels with confidence intervals, and threshold-selection diagnostics. I think the paper needs to name it and say clearly what nsEVDx adds over it. From what I can tell the real differentiator is first-class non-stationary covariates on location/scale/shape in one Python API covering both frequentist and Bayesian, which is a fair niche, but right now that has to be inferred rather than read.

Validation: the paper asserts the samplers and estimators work but doesn't show it. For a package whose value is the inference itself, it would help a lot to see parameter recovery on simulated data (stationary and non-stationary), and at least one cross-check of a stationary fit against scipy or extRemes so a reader can trust the numbers. I ran the frequentist path against a scipy-generated GEV and hit an API error before I could compare (filed separately), which is part of why this matters.

Validation: the paper asserts the samplers and estimators work but doesn't show it. For a package whose value is the inference itself, it would help a lot to see parameter recovery on simulated data (stationary and non-stationary), and at least one cross-check of a stationary fit against scipy or extRemes so a reader can trust the numbers. I ran the frequentist path against a scipy-generated GEV to do exactly this kind of cross-check and hit an API error before I could compare , passing the distribution as the string "gev" throws `AttributeError: 'str' object has no attribute 'name' in suggest_bounds (evd_model.py:237)`. That's the same drift  documented in Nischalcs50/nsEVDx#11, so I'm not filing it separately, but it's part of why an explicit validation example matters , the intended cross-check isn't currently runnable through the documented string API. Relatedly, it'd be worth a test that exercises the documented string API ("gev"), since that path currently throws on basic use and the suite doesn't seem to catch it.

Example: the one worked example fits essentially stationary Gumbel data, but non-stationarity is the headline feature. A single covariate-driven example end to end (fit, diagnostics, a model-comparison step) would show the core capability.

Smaller things: Coles is usually cited as 2001 for the EVD text, worth checking the year; kafle_detecting is cited as evidence of use but is listed as no-date/upcoming, so maybe mark it in prep; and a funding/no-funding statement is the JOSS convention.

To be clear on priority: the state-of-field comparison and a runnable non-stationary example feel like the things worth doing before acceptance; the citation and funding points are trivial cleanups. None of this is meant harshly the niche is real and the package is lightweight in a good way, and these are the things that would let a reader trust and adopt it.
