### 2026-08-25 Nischalcs50/nsEVDx#8 @Nischalcs50
@manunicholasjacob
Thank you for the detailed investigation and explanation.
I agree that there appears to be a mismatch between the adopted mass-matrix convention and what is described in Neal (2011). I am currently working to fix this issue.
Please share the benchmark script you used. It would be very useful for validating the revised implementation.

Best regards,
Nischal
