### 2026-08-27 Nischalcs50/nsEVDx#21 @manunicholasjacob
[JOSS Review] HMC tempering no longer changes the sampled posterior
Part of the JOSS review at openjournals/joss-reviews#11187.

`_hamiltonian` changed in 5f1102b from `(-log_post/T) + K` to `(-log_post + K)/T`. I think the first form was right. It went in with the GPD support fix, so it is easy to miss.

Momentum is drawn at hmc_engine.py:290 as `N(0, M)`, so its density is `exp(-K)` with K untempered. To target `p(θ)^(1/T)` you need `H = -log p/T + K`, so the momentum marginal of `exp(-H)` matches the draw. Dividing K by T as well implies `N(0, T·M)`, which is not what gets drawn.

On a standard Gaussian, `p^(1/T)` is exactly `N(0, T)`, so the sampled sd should be `sqrt(T)`. 20,000 draws, 20 leapfrog steps, M = I:

| T | expected | before 5f1102b | master |
|---|---|---|---|
| 2 | 1.414 | 1.367 | 0.988 |
| 4 | 2.000 | 1.878 | 0.992 |
| 10 | 3.162 | 2.495 | 0.995 |

The spread does not move with T. Acceptance climbs to 0.999 instead, so ΔH is being flattened rather than the target. I get the same ratios across step sizes 0.15, 0.35 and 0.6 and two seeds, 12 runs each way.

The docstring at lines 107-109 still says T scales the potential energy.

Nothing shipped is affected. The non-unit T values in the notebooks go to `MH_RandWalk` and `MH_Mala`, which never call `_hamiltonian`, and `test_MH_Hmc` uses T=1, so the suite stays green either way.

I can open a PR reverting the line and adding a test that the sampled variance tracks T.
