### 2026-08-27 Nischalcs50/nsEVDx#9 @manunicholasjacob
Verified on master at 4fc989b. This one looks done.

Both `neg_log_likelihood_ns` and `_grad_nll_gpd` reject `z < 0` now. Same reproduction, same 300 points: the NLL at the true parameters is still 844.545 and still matches scipy. With the location at 1.0, which puts 58 points below support, I get Safe_INF instead of the 783.55 I reported, and the gradient there is no longer finite. The random-walk fit recovers [-0.009, 5.194, 0.184] against truth [0, 5, 0.2], where before it drifted to [2.887, 0.368, 0.149].

`test_ns_gpd_rejects_data_below_location` and the gradient equivalent cover it.
