### 2026-08-27 Nischalcs50/nsEVDx#11 @manunicholasjacob
Verified on master at 4fc989b. All four look done.

A string dist works through both `frequentist_nsEVD` and `suggest_bounds`, and "genextreme", "gev" and the scipy object give identical bounds. The return contract lines up now, annotation and docstring both `np.ndarray` against a `res.x` return. `_hmc_step` describes what it actually returns, and the numerical gradient is renamed `_numerical_grad_log_likelihood` with a note that the prior part is added in `_grad_log_posterior`.

One leftover in that docstring: Returns still calls `params_new` "the parameter vector for the next state (either proposed or current)", but it returns `q_prop` unconditionally. Not worth its own issue.
