### 2026-09-01 a1d8f64
Changes in examples, README.md, paper.md, added Figure 3 to paper.md, minor changes in the source codes, minor change in toml file
