+++ Paper/paper.md
@@ -27,21 +27,22 @@ bibliography: paper.bib
 csl: apa.csl
 ---
 # Summary
-Infrastructure design, insurance pricing, and financial planning all rely on understanding the probability of rare, extreme events, such as extreme floods, rainfall events, and stock market crashes. Traditionally, researchers estimated the probability of extreme events of a given magnitude by assuming that their probability of occurrence remained stationary over time. However, changing global conditions such as climate change, urbanization, and economic shifts can cause the probability of extreme events to become non-stationary. `nsEVDx` is an open-source Python package developed to help researchers quantify these changing risks by fitting both stationary and non-stationary Extreme Value Distributions (EVDs) to extreme value data. It supports non-stationary modeling in which the location, scale, and shape parameters can each depend on user-defined covariates. This enables practical applications such as linking rainfall extremes to temperature or maximum stock market losses to market volatility indices. Overall, `nsEVDx` provides researchers and practitioners with an easy-to-use, flexible, and extensible Python-native interface for frequentist and Bayesian inference, including MCMC methods.
+
+`nsEVDx` is an open-source Python package for fitting stationary and non-stationary Extreme Value Distributions (EVDs) to extreme value data. It supports non-stationary models in which the location, scale, and shape parameters can each depend on user-defined covariates. Overall, `nsEVDx` provides an easy-to-use, flexible, and extensible Python-native interface for frequentist and Bayesian inference of non-stationary EVD models. This enables researchers and practitioners to quantify changing risks associated with rare and extreme events.
 
 
 # Statement of Need
 
 Traditional extreme value frequency analysis assumes that the statistical properties of extreme events remain stationary over time. However, non-stationary behavior in extreme events has been documented across hydroclimatic, engineering, and financial applications. For example, in hydroclimatic systems, non-stationary behavior has been documented in flood and extreme-rainfall processes, with changes linked to factors such as urbanization and climate change [@prosdocimi2015detection;@jayaweera2025evidence;@eccles2025substantial]. Such changes can alter the frequency and magnitude of extreme events over time. Accurately quantifying these evolving risks and return periods requires statistical tools that can incorporate arbitrary, time-varying covariates into all parameters of EVDs.
 
-While several software tools support extreme value analysis, implementing and evaluating non-stationary EVD models often requires substantial statistical and programming expertise, particularly when Bayesian inference and advanced sampling algorithms are involved. `nsEVDx` provides an easy-to-use framework for fitting stationary and non-stationary Generalized Extreme Value (GEV) and Generalized Pareto (GPD) models. It makes advanced sampling methods accessible to users with varying levels of expertise in extreme value modeling while retaining the flexibility required for advanced applications.
+Although several existing packages support extreme value analysis, modeling non-stationary EVDs requires substantial statistical and programming expertise, particularly when Bayesian sampling algorithms are involved. `nsEVDx` provides an easy-to-use framework for fitting stationary and non-stationary Generalized Extreme Value (GEV) and Generalized Pareto (GPD) distributions. It makes advanced sampling methods accessible to users with varying levels of expertise in extreme value modeling while retaining the flexibility required for advanced applications.
 
[…]
