+++ pyproject.toml
@@ -56,3 +56,6 @@ select = ["E", "F", "W", "I"] # Error, Flake8, Warnings, Isort
 [tool.pytest.ini_options]
 testpaths = ["tests"]
 addopts = "--cov=nsEVDx --cov-report=term-missing"
+
+[tool.setuptools.packages.find]
+where = ["nsEVDx"]
