### 2026-09-01 8b797cc
references added to README.md and some edits in the paper to add comparision with other packages
