+++ Paper/paper.md
@@ -28,43 +28,41 @@ csl: apa.csl
 ---
 # Summary
 
-`nsEVDx` is an open-source Python package for fitting stationary and non-stationary Extreme Value Distributions (EVDs) to extreme value data. It supports non-stationary models in which the location, scale, and shape parameters can each depend on user-defined covariates. Overall, `nsEVDx` provides an easy-to-use, flexible, and extensible Python-native interface for frequentist and Bayesian inference of non-stationary EVD models. This enables researchers and practitioners to quantify changing risks associated with rare and extreme events.
+nsEVDx is an open-source Python package for fitting stationary and non-stationary Extreme Value Distributions (EVDs) to extreme value data. It supports non-stationary models in which the location, scale, and shape parameters can each depend on user-defined covariates. Overall, nsEVDx provides a Python-native interface for frequentist and Bayesian inference of non-stationary EVD models, enabling researchers and practitioners to quantify changing risks associated with rare and extreme events.
 
 
 # Statement of Need
 
-Traditional extreme value frequency analysis assumes that the statistical properties of extreme events remain stationary over time. However, non-stationary behavior in extreme events has been documented across hydroclimatic, engineering, and financial applications. For example, in hydroclimatic systems, non-stationary behavior has been documented in flood and extreme-rainfall processes, with changes linked to factors such as urbanization and climate change [@prosdocimi2015detection;@jayaweera2025evidence;@eccles2025substantial]. Such changes can alter the frequency and magnitude of extreme events over time. Accurately quantifying these evolving risks and return periods requires statistical tools that can incorporate arbitrary, time-varying covariates into all parameters of EVDs.
+Traditional extreme value frequency analysis assumes that the statistical properties of extreme events remain stationary over time. However, non-stationary behavior in extreme events has been documented across hydroclimatic, engineering, and financial applications. For example, in hydroclimatic systems, non-stationary behavior has been documented in flood and extreme-rainfall processes, with changes linked to factors such as urbanization and climate change [@prosdocimi2015detection;@jayaweera2025evidence;@eccles2025substantial]. Such changes can alter the frequency and magnitude of extreme events over time. Accurately quantifying these evolving risks and return periods requires statistical tools that can incorporate arbitrary, time-varying covariates into all EVD parameters.
 
 Although several existing packages support extreme value analysis, modeling non-stationary EVDs requires substantial statistical and programming expertise, particularly when Bayesian sampling algorithms are involved. `nsEVDx` provides an easy-to-use framework for fitting stationary and non-stationary Generalized Extreme Value (GEV) and Generalized Pareto (GPD) distributions. It makes advanced sampling methods accessible to users with varying levels of expertise in extreme value modeling while retaining the flexibility required for advanced applications.
 
-The target audience for `nsEVDx` includes hydrologists, climate scientists, engineers, and risk analysts who require a reliable and flexible workflow for non-stationary extreme value modeling. By reducing the effort  needed to implement both frequentist optimization and advanced Markov Chain Monte Carlo (MCMC) sampling algorithms, the software allows practicioners to focus on understanding changing risks and their implications for decision-making.
+The target audience for nsEVDx includes hydrologists, climate scientists, engineers, and risk analysts who need a reliable, flexible workflow for non-stationary extreme value modeling. By reducing the effort needed to implement both frequentist optimization and advanced Markov Chain Monte Carlo (MCMC) sampling algorithms, the software allows practitioners to focus on understanding changing risks and their implications for decision-making.
 
 
 # State of the Field
 
[…]
