+++ Paper/paper.bib
@@ -220,7 +220,7 @@ Pages: 2.2-1}
 
 @article{kafle2026robustness,
   author = {Kafle, Nischal and Dell'Aira, Francesco and Chadwick, Cristián and Meier, Claudio I.},
-  title = {(Forthcoming.) Robustness of Regionally Derived, Short-Duration Rainfall Depth-Duration-Frequency Estimates to the Choice of Minimum Interevent Time: Evidence across Climates, Raingauge Densities, and Regionalization Approaches},
+  title = {Robustness of Regionally Derived, Short-Duration Rainfall Depth-Duration-Frequency Estimates to the Choice of Minimum Interevent Time: Evidence across Climates, Raingauge Densities, and Regionalization Approaches},
   journal = {Journal of Hydrologic Engineering},
   year = {2026},
   doi = {10.1061/JHYEFF.HEENG-6729}
@@ -243,8 +243,8 @@ Pages: 2.2-1}
 }
 
 @article{emcee,
-   author = {{Foreman-Mackey}, D. and {Hogg}, D.~W. and {Lang}, D. and {Goodman}, J.},
     title = {emcee: The MCMC Hammer},
+	author = {{Foreman-Mackey}, D. and {Hogg}, D.~W. and {Lang}, D. and {Goodman}, J.},
   journal = {PASP},
      year = 2013,
    volume = 125,
@@ -254,8 +254,8 @@ Pages: 2.2-1}
 }
 
 @incollection{neal2011mcmc,
+  title     = {MCMC Using Hamiltonian Dynamics},
   author    = {Neal, Radford M.},
-  title     = {{MCMC} Using {H}amiltonian Dynamics},
   booktitle = {Handbook of Markov Chain Monte Carlo},
   editor    = {Brooks, Steve and Gelman, Andrew and Jones, Galin L. and Meng, Xiao-Li},
   publisher = {CRC Press},
