+++ README.md
@@ -160,44 +160,51 @@ Please see [CONTRIBUTING.md](CONTRIBUTING.md) for guidelines on contributing to 
 
 ## References
 
-Abril-Pla, O., Andreani, V., Carroll, C., Dong, L. Y., Fonnesbeck, C., Kochurov, M., Kumar, R., Lao, J., Luhmann, C. C., Martin, O. A., Osthege, M., Vieira, R., Wiecki, T. V., & Zinkov, R. (2023). PyMC: A modern, and comprehensive probabilistic programming framework in Python. *PeerJ Computer Science, 9*, e1516. https://doi.org/10.7717/peerj-cs.1516
+Betancourt, M. (2017). A Conceptual Introduction to Hamiltonian Monte Carlo. arXiv: Methodology. https://doi.org/10.48550/arXiv.1701.02434
 
-Betancourt, M. (2017). *A conceptual introduction to Hamiltonian Monte Carlo*. arXiv. https://arxiv.org/abs/1701.02434
+Castillo, E. (1988). Extreme value theory in engineering. Academic Press. https://doi.org/10.1016/C2009-0-22169-6
 
-Castillo, E. (1988). *Extreme value theory in engineering*. Academic Press.
+Coles, S. (2001). An introduction to statistical modeling of extreme values (4th. printing). Springer. https://doi.org/10.1007/978-1-4471-3675-0
 
-Eccles, R., Syktus, J., Trancoso, R., Chapman, S., Wasko, C., Evans, J. P., Thatcher, M., Di Virgilio, G., & Stassen, C. (2025). Substantial increases in future precipitation extremes—Insights from a large ensemble of downscaled CMIP6 models. *npj Natural Hazards, 2*(1), 60. https://doi.org/10.1038/s44304-025-00107-1
+Deville, Y. (2026). NSGEV: Non-stationary GEV time series. https://github.com/IRSN/NSGEV/
 
-Gilleland, E. (2025). *extRemes: Extreme value analysis* (Version 2.2-1) [R package]. Comprehensive R Archive Network. https://CRAN.R-project.org/package=extRemes
+Eccles, R., Syktus, J., Trancoso, R., Chapman, S., Wasko, C., Evans, J. P., Thatcher, M., Di Virgilio, G., & Stassen, C. (2025). Substantial increases in future precipitation extremes—insights from a large ensemble of downscaled CMIP6 models. Npj Natural Hazards, 2(1), 60. https://doi.org/10.1038/s44304-025-00107-1
 
-Heffernan, J. E., Stephenson, A. G., & Gilleland, E. (2003). *ismev: An introduction to statistical modeling of extreme values* (Version 1.42) [R package]. Comprehensive R Archive Network. https://CRAN.R-project.org/package=ismev
+Foreman-Mackey, D., Hogg, D. W., Lang, D., & Goodman, J. (2013). Emcee: The MCMC hammer. PASP, 125, 306–312. https://doi.org/10.1086/670067
 
-Hosking, J. R. M., & Wallis, J. R. (1997). *Regional frequency analysis: An approach based on L-moments*. Cambridge University Press.
+Gilleland, E. (2025). extRemes: Extreme Value Analysis. https://doi.org/10.1175/JTECH-D-20-0070.1
 
-Jayaweera, L., Wasko, C., & Nathan, R. (2025). Evidence for non-stationarity in the GEV shape parameter when modeling extreme rainfall. *Water Resources Research, 61*(5), e2023WR036426. https://doi.org/10.1029/2023WR036426
+Heffernan J. E., Stephenson A.G., & Gilleland E. (2003). Ismev: An Introduction to Statistical Modeling of Extreme Values. https://doi.org/10.32614/CRAN.package.ismev
 
[…]
