### 2026-09-01 fb038ee
all (138) tests pass | all examples rerun successfully
