### 2026-09-01 f1563ca
Bump version to v0.2.5
