### 2026-09-01 75c66e2
Fix: Issue in HMCengine_test.py due to change in _hmc_step() return
