### 2026-09-01 17d5a9b
Bump version to v0.2.6
