+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = nsEVDx
-version = 0.2.5
+version = 0.2.6
 description = Modeling Non-stationary Extreme Value Distributions
 long_description = file: README.md
 long_description_content_type = text/markdown
