### 2026-09-01 6d41c56
Fix packaging and bump version to v0.2.7
