### 2026-09-01 a7cbdb0
added no funding to the paper
