### 2026-09-01 c207ecd
(auto) Paper PDF Draft
