### 2026-09-01 #11187 @Nischalcs50
I would like to sincerely thank the Editor @boisgera and both reviewers for their thorough review, extensive testing, and constructive feedback on `nsEVDx`. Their comments, bug reports, benchmark scripts, pull requests, and suggestions have significantly improved both the package and the accompanying manuscript.

All issues raised during the review process have now been addressed, and a new release (**v0.2.7**) has been published on `PyPI`. Please proceed with any further verification or review steps at your convenience, and I would be happy to address any additional questions or comments that may arise.

## Summary of Changes

### **Reviewer 1** @manunicholasjacob
I would also like to express my gratitude to Reviewer 1 for the detailed code review, regression tests, benchmark scripts, and pull requests submitted during the review process.

- **[Issue #6](https://github.com/Nischalcs50/nsEVDx/issues/6#) Sign error in the MALA acceptance ratio:** I have merged the [PR #14](https://github.com/Nischalcs50/nsEVDx/pull/14) created by Reviewer 1. The sign error in the MALA acceptance ratio has been corrected, and additional regression tests based on Gaussian targets with known analytical solutions were added to validate the implementation. The revised sampler was verified to recover the correct target distribution and posterior spread.

- **[Issue #7](https://github.com/Nischalcs50/nsEVDx/issues/7#) Reported acceptance rates include burn-in acceptances:** Thank you for submitting a PR to address this issue. I have merged the [PR #15](https://github.com/Nischalcs50/nsEVDx/pull/15/changes) created by Reviewer 1. The acceptance-rate calculations were updated so that they are reported using only production samples, excluding burn-in iterations.

- **[Issue #8](https://github.com/Nischalcs50/nsEVDx/issues/8#) HMC mass matrix appears to use the inverse convention:** Please see commit [87bae35](https://github.com/Nischalcs50/nsEVDx/commit/87bae35d569f50e1abfd7fc179ca9e73bd520969) for the changes made in hmc_engine.py to address this issue. The HMC mass-matrix adaptation has been updated to use the precision convention discussed in Neal (2011), and both the Hessian-based initialization and warmup adaptation procedures were modified accordingly. Implemented the warmup acceptance/rejection logic that was previously missing. Thank you for providing the benchmark scripts and opening [PR #20](https://github.com/Nischalcs50/nsEVDx/pull/20), which were extremely helpful in validating the revised implementation. The updated implementation was evaluated using the benchmark methodology discussed in the issue and showed substantial improvements in effective sample size and overall sampling performance.

- **[Issue #9](https://github.com/Nischalcs50/nsEVDx/issues/9) GPD likelihood does not enforce the lower support bound:** Thank you for identifying this issue. Please see commit [5f1102b](https://github.com/Nischalcs50/nsEVDx/commit/5f1102b0c5525fcedfbd052e610efb61d1850b8b) made in response to this issue. The missing lower-support condition (`z >= 0`) has been added to the GPD likelihood implementation, and the same support validation has been incorporated into the analytical gradient calculations. In addition, regression tests were added to ensure that future updates do not violate the theoretical support constraints of the GPD distribution.

- **[Issue #10](https://github.com/Nischalcs50/nsEVDx/issues/10) Released package differs materially from the current master branch:** A new release `v0.2.7` containing the latest changes is available at [`PiPy`](https://pypi.org/project/nsEVDx/). This release incorporates the fixes, bug corrections, documentation updates, and manuscript-related revisions that were previously available only on the development branch. The published package is now consistent with the current state of the repository.

- **[Issue #11](https://github.com/Nischalcs50/nsEVDx/issues/11) A few API contracts and docstrings have drifted from the implementation:** Please see commit [1c7bf89](https://github.com/Nischalcs50/nsEVDx/commit/1c7bf8946fcfc77b15880f03037f6c7cd4e3ddb9) to verify the changes made to address this issue. The docstrings, API contracts, examples, and documentation have been updated to match the current implementation. In addition, outdated references and example code were corrected, and inconsistencies in the Quick Start example were resolved to ensure that the documented behavior aligns with the software.

- **[Issue #12](https://github.com/Nischalcs50/nsEVDx/issues/12) Paper and reference cleanup:** I have merged the [PR #16](https://github.com/Nischalcs50/nsEVDx/pull/16) created by Reviewer 1 and made several additional revisions to the manuscript and references throughout the development process. Please refer to the latest version of the paper. The following changes have been implemented:
  - Updated references and corrected citation issues.
  - Added a comparison with the Python package `pyextremes` in the **State of the Field** section.
  - Added a **Benchmark Experiment** section demonstrating parameter recovery and agreement with known targets under stationary and non-stationary settings.
  - Added a short non-stationary modeling example to better illustrate the package's primary capability.
  - Corrected reference metadata and citation inconsistencies.

- **[Issue #13](https://github.com/Nischalcs50/nsEVDx/issues/13) Default priors and quick-start tempering may be misleading for new users:** Please see commits [`373a77b`](https://github.com/Nischalcs50/nsEVDx/commit/373a77b005cac9ad9c6d84f1ac4392a185c6fdde#diff-03d77dd52244d2e1c102ac844797b05f436d015a952173d6dfe4a76d12e1b49d), [`a1d8f64`](https://github.com/Nischalcs50/nsEVDx/commit/a1d8f649cc4a520b12a1a78382c50e9dc0970701#diff-03d77dd52244d2e1c102ac844797b05f436d015a952173d6dfe4a76d12e1b49d), and [`2ddf73f`](https://github.com/Nischalcs50/nsEVDx/commit/2ddf73f3097d66ed5432646f30df7f728c0da9fa) for the corresponding changes. The prior specification framework has been revised to provide a more general and data-driven approach. Prior distributions are now generated based on both the scale of the response variable and the scale of the covariates. In addition, a truncated normal prior is now used for the shape parameter of extreme value distributions to improve robustness and avoid unrealistic values. The Quick Start example and accompanying documentation have also been updated to better reflect typical usage patterns, clarify prior selection, and reduce potential confusion for new users.

### **Reviewer 2** @dosvk
I also thank Reviewer 2 for carefully testing the released package, verifying the examples, and identifying several packaging, documentation, and usability issues.

- **[Issue #17](https://github.com/Nischalcs50/nsEVDx/issues/17) README Quick Start fails on the released package (v0.2.4) MH_Hmc return signature mismatch:** A new release `v0.2.7` containing the latest changes is available at [`PiPy`](https://pypi.org/project/nsEVDx/). This version is free of this issue. Please re-run the README Quick Start Example.

- **[Issue #18](https://github.com/Nischalcs50/nsEVDx/issues/18) Plotting tests block on figure windows without a headless backend:** I have updated plotting utilities to support `show=False`. Plotting functions now return figure objects and no longer require interactive display during testing. Tests should now execute correctly without blocking on figure windows. Please see commit [55ca15b](https://github.com/Nischalcs50/nsEVDx/commit/55ca15bb0d85802f28e587f44f37808fe869f0a9) to verify the changes made to address this issue.

- **[Issue #19](https://github.com/Nischalcs50/nsEVDx/issues/19) Bare pytest fails on a fresh clone, pytest-cov missing from requirements.txt:** Please see commit [0b5248b](https://github.com/Nischalcs50/nsEVDx/commit/0b5248b5afed19ac26574c345740ec10041d3dfb), where `pytest-cov` was added to the development requirements. Thank you for identifying this issue. The test configuration has been updated accordingly, and the issue should now be resolved.

#### **Reviewer 2 comments related to the paper**

-  **Python EVA comparison is missing:** Thank you for this suggestion. I reviewed several existing Python libraries and, as suggested, identified `pyextremes` as one of the most relevant package. `pyextremes` has now been mentioned in the `State of the Field` section along with a discussion of the major differences between it and `nsEVDx`. In particular, `nsEVDx` supports non-stationary extreme value modeling with arbitrary covariates, whereas pyextremes primarily focuses on stationary workflows.

- **The paper has a Research Impact Statement but doesn't show the samplers/estimators recovering known values. For a package whose value is the inference itself, a small parameter-recovery on simulated data (stationary and non-stationary) plus one cross-check of a stationary fit against scipy or extRemes would let a reader trust the numbers:** Thank you for this valuable suggestion. I have added a new section **Benchmark Experiment**, to the paper. This section compares `nsEVDx` against `pyextremes`, 'SciPy`, and 'NumPyro`. While the benchmark is not intended to be a comprehensive validation study, as that would be beyond the scope of this manuscript, it demonstrates the ability of these packages to recover known extreme-value distribution parameters under both stationary and non-stationary settings. The comparison is based on the mean absolute error in recovering the true parameters across multiple simulations (`N = 100` for stationary models and `N = 50` for non-stationary models). The results are summarized in **Figure 3**.

- **Research impact / downloads: the statement cites 3,200+ first-year downloads. pepy currently shows ~3,500 total, so the number is fine, though a good chunk lands on release days, so I'd read it as early interest rather than sustained adoption. Reasonable to mention, just not something to lean on hard:** I have revised the Research Impact Statement and softened the language surrounding the download statistics. The statement now presents the download count as an indicator of early community interest, while keeping the focus on the package's intended scientific applications and potential impact.

- **The paper points to the repo/docs for examples rather than inlining one. That's a fair choice, but since non-stationarity is the headline feature, a short covariate-driven example somewhere the reader will see it, even in the linked docs, would show the core capability rather than leaving it to be inferred:** A covariate-driven example is added in the paper.

- **Smaller things: Coles is usually cited as 2001 for the EVD text, worth checking the year on coles_introduction_2007; and there's no funding/no-funding statement, which is the JOSS convention:** Thank you for pointing this out. The citation has been corrected, and a Funding section has been added to the manuscript following JOSS conventions. The manuscript now explicitly states that no external funding was received for this work.

Once again, I sincerely thank the Editor and both reviewers for their time, expertise, detailed testing, and constructive feedback. The review process has substantially improved the robustness, documentation, and usability of `nsEVDx`.
