### 2026-09-02 #11187 @dosvk
Thanks @Nischalcs50, re-verified against v0.2.7 from PyPI in a fresh Python 3.12 environment. My three findings are resolved.

The README Quick Start (F1) now runs verbatim end to end: MH_Hmc returns the results dict and the acceptance rates, sample means and R-hats print without the earlier unpacking error. The plotting tests (F3) call plot_trace with show=False, so they no longer block on a figure window in a headless run. And pytest-cov is now a declared dependency (F4), so the documented test path installs and runs, 139 passed at 97% coverage on my side.

One small note on F4, not a blocker: because addopts still forces --cov=nsEVDx, a bare `pytest` on a clone without the test extras installed still errors on the unknown argument. If you intend bare pytest to work for a casual contributor, moving --cov into a coverage-specific invocation would do it; otherwise a line in the contributing docs pointing at the test extra is enough.

I also confirmed the tests genuinely exercise the numerics rather than just running: flipping the sign on the core negative-log-likelihood return turns two tests red, which is exactly what you want.

On the paper side, everything I raised is addressed. pyextremes is now discussed in State of the Field with the stationary versus non-stationary distinction drawn clearly. The new Benchmark Experiment section is exactly the kind of validation I was hoping for, and going to a parameter-recovery comparison against SciPy, pyextremes and NumPyro across both stationary and non-stationary settings is more than I asked. The covariate-driven example is in, the download language is appropriately softened, and the Coles citation year is corrected.

With that, my checklist is complete, and from my side I'm happy to recommend acceptance, @boisgera.

Nice work turning all of this around, the package came out of the review meaningfully stronger.
