### 2026-09-02 #11187 @manunicholasjacob
Re-verified against v0.2.7 from PyPI in a fresh Python 3.12 environment, and against a clean clone at c207ecd. Both of my remaining boxes are ticked.

Issues #6 to #13 all confirmed resolved: the MALA forward density is now subtracted, accepts are counted only after burn-in, M_diag is the precision convention, z < 0 is rejected in both the GPD likelihood and its gradient, v0.2.7 on PyPI is byte-identical to master, docstrings match the implementation, prior scales are derived from the data with a truncated normal on the shape, and the paper's references are corrected.

On #8, I re-ran the benchmark. The two conventions still separate cleanly, and the revised code is on the right side of it: against a target spread of 50, the variance convention recovers 10.63, where precision recovers 49.10, and against a target of 100, it recovers 5.39, where precision recovers 98.20. End-to-end through the package's own adaptation, median minimum ESS is 214 against 11.7 before the change.

Functionality: 139 tests pass on a fresh clone with bare pytest, 97% coverage, and the README Quick Start runs verbatim, recovering [30, 0.1, 5, -0.3] as [28.9, 0.094, 4.57, -0.337] with all R-hat at 1.000.

Quality of writing: the State of the Field now separates the R packages, the general PPLs and pyextremes, and the Benchmark Experiment section answers whether the samplers recover known parameters rather than merely running.

I recommend acceptance. Congratulations, @Nischalcs50. The MALA sign error and the mass-matrix convention were both easy to overlook, since neither stops a chain from running and returning plausible numbers, and you fixed them properly rather than papering over them.
