### 2026-09-02 Nischalcs50/nsEVDx#8 @manunicholasjacob
Verified on v0.2.7. `_init_mass_matrix` now sets `M_diag` to the curvature, so M scales with posterior precision. Re-ran the benchmark rather than reading the diff: against a target spread of 50 the variance convention recovers 10.63 where precision recovers 49.10, and against 100 it recovers 5.39 where precision recovers 98.20. End to end, median minimum ESS is 214 against 11.7 before. Closing.
