### 2026-09-02 Nischalcs50/nsEVDx#9 @manunicholasjacob
Verified on v0.2.7. `z < 0` is now rejected in `neg_log_likelihood_ns`, which returns `Safe_INF`, and the same guard is in `_grad_nll_gpd`, so the analytical gradient cannot report a finite value outside the support either. Closing.
