### 2026-09-02 Nischalcs50/nsEVDx#10 @manunicholasjacob
Verified. `diff -rq` between the v0.2.7 install from PyPI and a clean clone at c207ecd reports no differences, so the released package and master no longer diverge. Closing.
