### 2026-09-02 Nischalcs50/nsEVDx#11 @manunicholasjacob
Verified on v0.2.7. Docstrings and examples match the implementation, and the README Quick Start runs verbatim end to end. Closing.
