### 2026-09-02 Nischalcs50/nsEVDx#12 @manunicholasjacob
Verified. Coles is now cited as 2001, `pyextremes` is covered in the State of the Field, and a Funding statement has been added. Closing.
