### 2026-09-02 Nischalcs50/nsEVDx#13 @manunicholasjacob
Verified on v0.2.7. Prior scales are now derived from the data rather than fixed, and the shape parameter uses a truncated normal bounded to [-0.5, 0.5]. Closing.
