### 2026-09-02 d49d89b
Revise CONTRIBUTING.md for clarity and dependency requirement
Updated contributing guidelines
