### 2026-09-02 e932062
added v0.2.7 zenedo doi to README
