### 2026-09-02 d55093e
v0.2.7 zenedo badge corrected
