### 2026-09-04 #11187 @boisgera
Hi @Nischalcs50,

Could you integrate https://github.com/Nischalcs50/nsEVDx/pull/24 then perform the actions listed in the **Additional Author Tasks After Review is Complete** section above? Please make sure that the title your DOI archive is exactly the one of your JOSS paper ("nsEVDx: A Python Library for Stationary and Non-Stationary Extreme Value Modeling").

Best,

Sébastien
