+++ setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
 name = nsEVDx
-version = 0.2.7
-description = Modeling Non-stationary Extreme Value Distributions
+version = 0.2.8
+description = Modeling Stationary and Non-stationary Extreme Value Distributions
 long_description = file: README.md
 long_description_content_type = text/markdown
 author = Nischal Kafle
