### 2026-09-04 #11187 @Nischalcs50
Hi Professor @boisgera,

Thank you for your guidance and support throughout the review process.

I have now completed the post-review author tasks:

- Merged PR 24 and created a new software release. The new release version is v0.2.8.
- Archived the release on Zenodo. The DOI is: https://doi.org/10.5281/zenodo.22307685
- Verified that the title, author information (including ORCIDs), and license are consistent between the JOSS paper and the archived release. The archive title matches the manuscript title exactly: "nsEVDx: A Python Library for Stationary and Non-Stationary Extreme Value Modeling".

Thank you again for your time, support, and careful handling of this review process. I greatly appreciate all the effort you and the reviewers have invested in improving the manuscript and software.

Best regards,
Nischal Kafle
