### 2026-09-05 #11187 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
