### 2026-09-05 #11187 @crvernon
@Nischalcs50 - I just need you to address the following before I accept this submission for publication:

- [ ] LINE 46: The citation "Oriol Abril-Pla et al. (2023)" appears to result from incorrectly parsed author names in the bibliography. It should be "Abril-Pla et al. (2023)".
- [ ] LINES 53-57: pyextremes is discussed but has no corresponding citation/reference. Add an appropriate software or publication citation.
- [ ] LINES 60-62: Introduce the sampler acronyms with parentheses rather than commas. For example, use "Metropolis-adjusted Langevin algorithm (MALA)" and "Hamiltonian Monte Carlo (HMC)".
- [ ] LINE 138: The code comment "# Infering GEV Parameters from the Observations" contains a typo. "Infering" should be "Inferring".
- [ ] LINES 142-144: The AI Usage Disclosure is too general about how AI was used. JOSS asks authors who used generative AI to describe its use and how generated material was checked. Specify the kinds of development/manuscript tasks for which AI assistance was used and the verification process.
- [ ] LINES 167-168: The DOI given for the extRemes software reference appears inconsistent with a software-package citation. Verify that the DOI actually corresponds to the cited extRemes package/version rather than to a different publication.
- [ ] LINES 195-197: In the Neal reference, "hamiltonian dynamics" should preserve the proper technical capitalization "Hamiltonian dynamics".
- [ ] LINES 195-197: "Handbook of markov chain monte carlo" should capitalize the proper technical name consistently: "Handbook of Markov Chain Monte Carlo".
- [ ] LINES 217-218: "PyStan: The python interface to Stan" should capitalize the programming language: "PyStan: The Python interface to Stan."

Let me know when these have been addressed.  Thank you.
