### 2026-09-05 570fb5d
Fixed ref to Abril-Pla L46 of paper
