+++ Paper/paper.bib
@@ -3,9 +3,21 @@
 	title = {{PyMC}: a modern, and comprehensive probabilistic programming framework in {Python}},
 	volume = {9},
 	doi = {10.7717/peerj-cs.1516},
-	abstract = {PyMC is a probabilistic programming library for Python that provides tools for constructing and fitting Bayesian models. It offers an intuitive, readable syntax that is close to the natural syntax statisticians use to describe models. PyMC leverages the symbolic computation library PyTensor, allowing it to be compiled into a variety of computational backends, such as C, JAX, and Numba, which in turn offer access to different computational architectures including CPU, GPU, and TPU. Being a general modeling framework, PyMC supports a variety of models including generalized hierarchical linear regression and classification, time series, ordinary differential equations (ODEs), and non-parametric models such as Gaussian processes (GPs). We demonstrate PyMC's versatility and ease of use with examples spanning a range of common statistical models. Additionally, we discuss the positive role of PyMC in the development of the open-source ecosystem for probabilistic programming.},
 	journal = {PeerJ Computer Science},
-	author = {{Oriol Abril-Pla} and {Virgile Andreani} and {C. Carroll} and {L. Y. Dong} and {Christopher Fonnesbeck} and {Maxim Kochurov} and {Ravin Kumar} and {Junpeng Lao} and {Christian C. Luhmann} and {Osvaldo A. Martin} and {Michael Osthege} and {Ricardo Vieira} and {Thomas V. Wiecki} and {Robert Zinkov}},
+	author = {Abril-Pla, Oriol and
+          Andreani, Virgile and
+          Carroll, C. and
+          Dong, L. Y. and
+          Fonnesbeck, Christopher and
+          Kochurov, Maxim and
+          Kumar, Ravin and
+          Lao, Junpeng and
+          Luhmann, Christian C. and
+          Martin, Osvaldo A. and
+          Osthege, Michael and
+          Vieira, Ricardo and
+          Wiecki, Thomas V. and
+          Zinkov, Robert}
 	month = {sep},
 	year = {2023},
 	doi = {10.7717/peerj-cs.1516},
