### 2026-09-05 4bfe5e0
typos fixed L195-197 in paper
