+++ Paper/paper.bib
@@ -17,7 +17,7 @@
           Osthege, Michael and
           Vieira, Ricardo and
           Wiecki, Thomas V. and
-          Zinkov, Robert}
+          Zinkov, Robert},
 	month = {sep},
 	year = {2023},
 	doi = {10.7717/peerj-cs.1516},
@@ -266,9 +266,9 @@ Pages: 2.2-1}
 }
 
 @incollection{neal2011mcmc,
-  title     = {MCMC Using Hamiltonian Dynamics},
+  title     = {MCMC Using {H}amiltonian {D}ynamics},
   author    = {Neal, Radford M.},
-  booktitle = {Handbook of Markov Chain Monte Carlo},
+  booktitle = {Handbook of {M}arkov {C}hain {M}onte {C}arlo},
   editor    = {Brooks, Steve and Gelman, Andrew and Jones, Galin L. and Meng, Xiao-Li},
   publisher = {CRC Press},
   address   = {Boca Raton, FL},
