### 2026-09-05 74ab0a1
fixed doi of extRemes L167-168
