+++ Paper/paper.bib
@@ -119,7 +119,7 @@ Pages: 0.3.1},
 	title = {{extRemes}: {Extreme} {Value} {Analysis}},
 	shorttitle = {{extRemes}},
 	url = {https://CRAN.R-project.org/package=extRemes},
-	doi = {10.1175/JTECH-D-20-0070.1},
+	doi = {https://doi.org/10.32614/CRAN.package.extRemes},
 	abstract = {General functions for performing extreme value analysis.  In particular, allows for inclusion of covariates into the parameters of the extreme-value distributions, as well as estimation through MLE, L-moments, generalized (penalized) MLE (GMLE), as well as Bayes.  Inference methods include parametric normal approximation, profile-likelihood, Bayes, and bootstrapping.  Some bivariate functionality and dependence checking (e.g., auto-tail dependence function plot, extremal index estimation) is also included.  For a tutorial, see Gilleland and Katz (2016) {\textless}doi: 10.18637/jss.v072.i08{\textgreater} and for bootstrapping, please see Gilleland (2020) {\textless}doi: 10.1175/JTECH-D-20-0070.1{\textgreater}.},
 	language = {en},
 	urldate = {2025-07-01},
