### 2026-09-05 7a48ac6
Fix: typos and issues in paper.md L60-62; L138; L142-144 (more specific AI usage)
