### 2026-09-05 c1099ef
updated zenedo doi at the paper metadata
