### 2026-09-05 945ec3d
corrected a ref related to PyStan L217-218
