+++ Paper/paper.bib
@@ -65,7 +65,7 @@ Pages: 0.3.1},
 }
 
 @misc{noauthor_pystan_2023,
-	title = {{PyStan}: {The} python interface to {Stan}},
+	title = {{PyStan}: {The} {Python} interface to {Stan}},
 	url = {https://pystan.readthedocs.io/},
 	author = {Stan Development Team,},
 	year = {2023},
