### 2026-09-05 dd9c76f
add bib ref to pyextremes
