+++ Paper/paper.bib
@@ -279,3 +279,13 @@ Pages: 2.2-1}
   archiveprefix = {arXiv},
   primaryclass  = {stat.CO}
 }
+
+@misc{bocharov_pyextremes,
+  author       = {Bocharov, George},
+  title        = {pyextremes: {E}xtreme {V}alue {A}nalysis (EVA) in {P}ython},
+  year         = {2020},
+  publisher    = {GitHub},
+  journal      = {GitHub repository},
+  url = {https://georgebv.github.io/pyextremes/},
+  version      = {2.5.0}
+}
