### 2026-09-05 #11187 @Nischalcs50
Thank you @crvernon for the detailed review.

I would also like to thank Professor Boisgérault for overseeing and completing the review process.

I have addressed all of the requested revisions:

- [x] **LINE 46:** I have corrected the bibliography entry so that the citation now appears as **"Abril-Pla et al. (2023)"** instead of **"Oriol Abril-Pla et al. (2023)"**.
- [x] **LINES 53-57:** Added citations to the **pyextremes** documentation and GitHub repository.
- [x] **LINES 60-62:** I have revised the sampler acronyms using parentheses, e.g., **Metropolis-adjusted Langevin algorithm (MALA)** and **Hamiltonian Monte Carlo (HMC)**. However, because the associated citations also used parentheses, placing them immediately afterward resulted in consecutive parentheses. To improve readability, the citations are now provided together at the end of the sentence.
- [x] **LINE 138:** Corrected the typo **"Infering"** to **"Inferring"** in the code comment.
- [x] **LINES 142-144:** Expanded the AI Usage Disclosure to explicitly describe how AI tools were used and how AI-generated content and code suggestions were independently reviewed, tested, verified, and revised before incorporation.
- [x] **LINES 167-168:** Verified and corrected the DOI link to **extRemes**.
- [x] **LINES 195-197:** Updated **"hamiltonian dynamics"** to **"Hamiltonian dynamics"**.
- [x] **LINES 195-197:** Updated **"Handbook of markov chain monte carlo"** to **"Handbook of Markov Chain Monte Carlo"**.
- [x] **LINES 217-218:** Corrected **"PyStan: The python interface to Stan"** to **"PyStan: The Python interface to Stan"**.

All requested changes have now been incorporated into the manuscript and references.

Please let me know if anything remains to be addressed or if any further revisions are required.
Thank you again for your time and careful review.

Nischal Kafle
