### 2026-09-05 739bb09
added JOSS DOI to README
